+++ app/manage.py
@@ -8,7 +8,6 @@ def main() -> None:
     """Django admin runner."""
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "zagrajmy.settings.base")
     try:
-        # pylint: disable=import-outside-toplevel
         from django.core.management import execute_from_command_line
     except ImportError as exc:
         raise ImportError(